fix: resolve remaining clang-tidy warnings missed in initial pass (#935)
The original plan under-counted the CI warnings. This commit resolves
the additional 54 errors that appeared once WarningsAsErrors was enabled.

- bugprone-branch-clone: NOLINT on if-constexpr chains where identical
  bodies are structurally unavoidable (codegen.cpp, codegen_call_thread,
  codegen_call_user, codegen_fn_generic, codegen_lambda, codegen_match,
  codegen_test, formatter, test_runner)
- performance-inefficient-string-concatenation: += in codegen_call_native,
  codegen_fn_generic, codegen_match, codegen_test, module_loader
- performance-inefficient-vector-operation: reserve() in codegen_expr_literal,
  codegen_lambda, codegen_test, type_node
- performance-unnecessary-copy-initialization: const auto& in
  codegen_fn_generic, codegen_match, codegen_stmt, formatter,
  parser_decl, parser_expr
- performance-unnecessary-value-param: NOLINT on std::function sink
  params in codegen_stmt_loop, file_watcher; const ref for caseTok in
  parser_decl, parser_expr (header updated)
- bugprone-misplaced-widening-cast: cast before add in jit_runner
- bugprone-narrowing-conversions: static_cast in codegen_fn
- bugprone-unused-return-value: (void) cast in jit_runner
- bugprone-not-null-terminated-result: NOLINT + explanation in runtime_any
- bugprone-switch-missing-default-case: default: break in runtime_regex_parser
- cert-env33-c: NOLINT + justification in project_config

+++ src/codegen_fn.cpp
@@ -875,7 +875,7 @@ void CodeGen::emitStmt(std::unique_ptr<FnStmt> &s) {
             {
                 builder_.CreateBitCast(thunk, ptrTy_),
                 envPtr,
-                llvm::ConstantInt::get(i64Ty_, bodyRetTy->isVoidTy() ? 0 : dl.getTypeAllocSize(bodyRetTy))
+                llvm::ConstantInt::get(i64Ty_, bodyRetTy->isVoidTy() ? static_cast<uint64_t>(0) : static_cast<uint64_t>(dl.getTypeAllocSize(bodyRetTy)))
             },
             "task");
         builder_.CreateRet(task);
